Cup 98' Plus name variant
This is for the current no-intro name.  It's unclear to me whether this name replaces the old name, or if they're different, though no-intro doesn't have the old name in database anymore.  Note this is a case where apparently the same ROM was published in multiple regions but with the same box art, based on the info I'm seeing.  I happened to choose the Japan box for this extended name variant.